pyproject.toml
[tool.poetry]
name = "Science-Fair-Website"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Flask = "^2.0.2"
dnspython = "^2.1.0"
pymongo = "^3.12.1"
python-dotenv = "^0.20.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
